+++ R/ftp-peptides.R
@@ -1,38 +1,28 @@
 #' Peptides
 #'
-#' @note Updated 2020-01-26.
+#' @note Updated 2021-03-12.
 #' @export
 #'
 #' @inheritParams params
 #' @inheritParams AcidRoxygen::params
 #'
 #' @return `SplitDataFrameList`.
-#' Split by `geneID` column.
+#' Split by `geneId` column.
 #'
 #' @examples
-#' ## WormBase FTP server must be accessible.
-#' tryCatch(
-#'     expr = peptides(),
-#'     error = function(e) e
-#' )
-peptides <- function(
-    version = NULL,
-    BPPARAM = BiocParallel::bpparam()  # nolint
-) {
-    file <- .assemblyFile(pattern = "wormpep_package", version = version)
+#' x <- peptides()
+#' print(x)
+peptides <- function(release = NULL) {
+    file <- .assemblyFile(stem = "wormpep_package.tar.gz", release = release)
     tempdir <- tempdir()
     ## Grep the verion number.
-    versionNumber <- str_match(file, "WS([[:digit:]]{3})")[1L, 2L]
+    releaseNumber <- str_match(file, "WS([[:digit:]]{3})")[1L, 2L]
     ## Extract the individual table.
-    wormpepTable <- paste0("wormpep.table", versionNumber)
-    untar(
-        tarfile = file,
-        files = wormpepTable,
-        exdir = tempdir
-    )
+    wormpepTable <- paste0("wormpep.table", releaseNumber)
+    status <- untar(tarfile = file, files = wormpepTable, exdir = tempdir)
+    assert(identical(status, 0L))
     x <- import(file = file.path(tempdir, wormpepTable), format = "lines")
-    cli_alert("Processing peptides.")
-    x <- bplapply(
+    x <- lapply(
         X = x,
         FUN = function(x) {
             sequence <- str_match(x, "^>([A-Za-z0-9\\.]+)")[[2L]]
@@ -45,20 +35,21 @@ peptides <- function(
             out <- c(pairs[, 3L])
             names(out) <- pairs[, 2L]
             out[["sequence"]] <- sequence
-            as.data.frame(t(out), stringsAsFactors = TRUE)
-        },
-        BPPARAM = BPPARAM
+            out
+        }
     )
-    x <- rbindlist(x, fill = TRUE)
-    x <- as(x, "DataFrame")
-    colnames(x)[colnames(x) == "gene"] <- "geneID"
-    x <- x[, unique(c("geneID", colnames(x)))]
-    keep <- grepl(pattern = genePattern, x = x[["geneID"]])
+    ## This step may need to be optimized.
+    x <- rbindToDataFrame(x)
+    colnames(x)[colnames(x) == "gene"] <- "geneId"
+    x <- x[, unique(c("geneId", colnames(x)))]
+    keep <- grepl(pattern = .genePattern, x = x[["geneId"]])
     x <- x[keep, , drop = FALSE]
     x <- x[
-        order(x[["geneID"]], x[["sequence"]], x[["wormpep"]]), , drop = FALSE
+        order(x[["geneId"]], x[["sequence"]], x[["wormpep"]]), , drop = FALSE
     ]
-    split(x, f = x[["geneID"]])
+    x <- split(x, f = x[["geneId"]])
+    assert(is(x, "SplitDataFrameList"))
+    x
 }
 
-formals(peptides)[["version"]] <- versionArg
+formals(peptides)[["release"]] <- .releaseArg