+++ tests/testthat/helper-globals.R
@@ -1,3 +1,6 @@
+release <- "WS279"
+ngene <- 52089L
+
 ## Ensure these aren't sorted by ID to check internal code.
 genes <- c(
     "skn-1" = "WBGene00004804",
@@ -5,8 +8,3 @@ genes <- c(
     "daf-2" = "WBGene00000898",
     "daf-16" = "WBGene00000912"
 )
-
-## Version-specific checks.
-## WS260: 51031.
-## WS269
-ngene <- 51254L