+++ tests/testthat/test-ftp-blastp.R
@@ -1,12 +1,20 @@
 context("ftp : blastp")
 
 test_that("current", {
-    x <- blastp(version = NULL)
+    x <- blastp(release = NULL)
     expect_s4_class(x, "SplitDataFrameList")
 })
 
 test_that("stable", {
-    x <- blastp(version = version)
+    x <- blastp(release = release)
     expect_s4_class(x, "SplitDataFrameList")
-    expect_identical(length(x), 14557L)
+    expect_identical(length(x), 14620L)
+    expect_identical(
+        object = x[[1L]],
+        expected = DataFrame(
+            "wormpep" = "CE00004",
+            "peptide" = "ENSP00000308897",
+            "eValue" = 1e-46
+        )
+    )
 })