+++ tests/testthat/test-ftp-geneOtherIDs.R
@@ -1,22 +1,19 @@
 context("ftp : geneOtherIDs")
 
 test_that("current", {
-    x <- geneOtherIDs(version = NULL)
-    expect_s4_class(x, "DataFrame")
+    x <- geneOtherIDs(release = NULL)
+    expect_s4_class(x, "CharacterList")
 })
 
 test_that("stable", {
-    x <- geneOtherIDs(version = version)
-    expect_s4_class(x, "DataFrame")
+    x <- geneOtherIDs(release = release)
+    expect_s4_class(x, "CharacterList")
     expect_identical(
-        object = lapply(x, class),
-        expected = list(
-            geneID = "character",
-            geneOtherIDs = "list"
-        )
+        object = length(x),
+        expected = ngene
     )
     expect_identical(
-        object = nrow(x),
-        expected = ngene
+        object = x[[1L]],
+        expected = c("aap-1", "Y110A7A.10")
     )
 })