+++ tests/testthat/test-rest-geneOntology.R
@@ -2,22 +2,24 @@ context("rest : geneOntology")
 
 test_that("current", {
     x <- geneOntology(genes)
-    expect_s4_class(x, "DataFrame")
-    expect_identical(x[["geneID"]], as.character(genes))
+    expect_s4_class(x, "SimpleList")
+    expect_identical(names(x), unname(genes))
     expect_identical(
-        lapply(x, class),
-        list(
-            geneID = "character",
-            biologicalProcess = "list",
-            cellularComponent = "list",
-            molecularFunction = "list"
+        object = names(x[[1L]]),
+        expected = c(
+            "biologicalProcess",
+            "cellularComponent",
+            "molecularFunction"
         )
     )
 })
 
 test_that("NULL return", {
+    genes <- "WBGene00000000"
+    x <- geneOntology(genes)
+    expect_identical(names(x), unname(genes))
     expect_identical(
-        geneOntology("WBGene00000000"),
-        NULL
+        object = x[[1L]],
+        expected = NULL
     )
 })